requirements.txt
Django==3.1.4
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
Pillow==8.0.1
six==1.14.0
django-ckeditor==6.0.0
django-import-export==2.5.0
